fix(bindings): stop iOS aborting on a negative DORS config number
The previous commit swept this bridge for conversions that trap on
JavaScript input, fixed thirteen of them, and added a guard that
fails on any `UInt8(...)` whose argument is not bounded where it is
written.

It missed two, and the guard could not have caught them, because
they are not byte conversions.

Both DORS config paths clamp `historyWindowSize` with
`max(1, min(100, Int(raw)))`, which reads as bounded and is not: the
narrowing runs *first*. `raw` is a `UInt64` straight off
`NSNumber.uint64Value`, and a negative JavaScript number lands there
as `UInt64.max`, so the `Int` conversion traps and aborts the
process before the clamp around it ever runs. A
`historyWindowSize` of -1 kills the app through `create()`, and
again through `updateDorsConfig`.

So clamp in the domain the value arrives in. `historyWindow` is only
ever consumed as a `UInt64`, so dropping the round-trip through
`Int` deletes the trapping conversion outright instead of making a
pointless one safe, and the two `UInt64(historyWindow)` call sites
become identity conversions and go with it. The symlink-farm
typecheck agrees it now infers as `UInt64`; the negative control
proved the harness by complaining about exactly that type.

No new guard for this one. Whether `Int(raw)` is safe depends on
what `raw` already is, which the text cannot say, and this file is
full of legitimate widenings like `Int(hops)` that a textual rule
would flag. Review and the checklist hold it instead, and both
bridge docs now say that a clamp wrapped *around* a narrowing
conversion does not count.

While at it, two comments on the new guards were simply wrong. The
byte guard's empty-argument skip claimed to exempt `[UInt8](...)`,
which never reaches it at all because the `]` stops it matching the
search; what it actually exempts is the zero-argument `UInt8()`. And
the ABI classifier claimed `Int?` boxes into an `NSNumber`. It
cannot: `@objc` refuses an optional value-type parameter outright,
so only reference types ever reach that branch.

+++ bindings/react-native/ios/OfflineProtocolModule.swift
@@ -608,7 +608,11 @@ class OfflineProtocolModule: RCTEventEmitter {
         let historyWindowRaw = UInt64((dorsDict["historyWindowSize"] as? NSNumber)?.uint64Value
                                       ?? (dorsDict["history_window_size"] as? NSNumber)?.uint64Value
                                       ?? 10)
-        let historyWindow = max(1, min(100, Int(historyWindowRaw)))
+        // Clamped in the unsigned domain, never through `Int`: a negative
+        // `historyWindowSize` from JavaScript reaches `uint64Value` as
+        // `UInt64.max`, and narrowing that to `Int` traps and aborts the
+        // application before the clamp around it can run.
+        let historyWindow = max(1, min(100, historyWindowRaw))
         let rawQueueRecovery = Float((dorsDict["queueRecoveryRatio"] as? NSNumber)?.floatValue
                                      ?? (dorsDict["queue_recovery_ratio"] as? NSNumber)?.floatValue
                                      ?? 0.5)
@@ -637,7 +641,7 @@ class OfflineProtocolModule: RCTEventEmitter {
             ttlEscalationThreshold: ttlThreshold,
             congestionDurationSecs: congestionDuration,
             ttlEscalationHoldSecs: ttlHold,
-            historyWindowSize: UInt64(historyWindow),
+            historyWindowSize: historyWindow,
             queueRecoveryRatio: queueRecovery,
             lowBatteryThreshold: lowBattery,
             relayMinBatteryLevel: relayMinBattery,
@@ -3316,7 +3320,11 @@ class OfflineProtocolModule: RCTEventEmitter {
             let congestionDuration = max((config["congestionDurationSecs"] as? NSNumber)?.uint64Value ?? current.congestionDurationSecs, 0)
             let ttlHold = max((config["ttlEscalationHoldSecs"] as? NSNumber)?.uint64Value ?? current.ttlEscalationHoldSecs, 1)
             let historyWindowRaw = (config["historyWindowSize"] as? NSNumber)?.uint64Value ?? current.historyWindowSize
-            let historyWindow = max(1, min(100, Int(historyWindowRaw)))
+            // Clamped in the unsigned domain, never through `Int`: a negative
+            // `historyWindowSize` from JavaScript reaches `uint64Value` as
+            // `UInt64.max`, and narrowing that to `Int` traps and aborts the
+            // application before the clamp around it can run.
+            let historyWindow = max(1, min(100, historyWindowRaw))
             let rawQueueRecovery = (config["queueRecoveryRatio"] as? NSNumber)?.floatValue ?? current.queueRecoveryRatio
             let queueRecovery = min(max(rawQueueRecovery, 0.0), 1.0)
 
@@ -3337,7 +3345,7 @@ class OfflineProtocolModule: RCTEventEmitter {
                 ttlEscalationThreshold: ttlThreshold,
                 congestionDurationSecs: UInt64(congestionDuration),
                 ttlEscalationHoldSecs: UInt64(ttlHold),
-                historyWindowSize: UInt64(historyWindow),
+                historyWindowSize: historyWindow,
                 queueRecoveryRatio: queueRecovery,
                 lowBatteryThreshold: UInt8(min(100, max(0, (config["lowBatteryThreshold"] as? NSNumber)?.intValue ?? Int(current.lowBatteryThreshold)))),
                 relayMinBatteryLevel: UInt8(min(100, max(0, (config["relayMinBatteryLevel"] as? NSNumber)?.intValue ?? Int(current.relayMinBatteryLevel)))),
